Fixing tests, improving the backend code
Changed config.py , improved conftest.py in middleware and User testing by switching to MongoClient. Changed test_data middleware and User. For the registered_user_in_db fixture, the logic of creating a random email was added, and for some tests, a random email using a uuid is also used